# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Client Wealth Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients/{clientId}/liabilities:
    get:
      operationId: listClientLiabilities
      tags: [liabilities]
      summary: List all liabilities for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of liabilities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Liability"
    post:
      operationId: createClientLiability
      tags: [liabilities]
      summary: Add a liability for a client (admin only)
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLiabilityBody"
      responses:
        "201":
          description: Liability created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Liability"
  /clients/{clientId}/liabilities/{liabilityId}:
    put:
      operationId: updateClientLiability
      tags: [liabilities]
      summary: Update a liability (admin only)
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: liabilityId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLiabilityBody"
      responses:
        "200":
          description: Updated liability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Liability"
    delete:
      operationId: deleteClientLiability
      tags: [liabilities]
      summary: Delete a liability (admin only)
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: liabilityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Liability deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    LoanType:
      type: string
      enum:
        - home_loan
        - car_loan
        - personal_loan
        - education_loan
        - business_loan
        - other
    CreateLiabilityBody:
      type: object
      properties:
        loanType:
          $ref: "#/components/schemas/LoanType"
        lenderName:
          type: string
        totalLoanAmount:
          type: number
        outstandingAmount:
          type: number
        interestRate:
          type: number
        emi:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        familyMemberId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - loanType
        - lenderName
        - totalLoanAmount
        - outstandingAmount
        - interestRate
        - emi
    UpdateLiabilityBody:
      type: object
      properties:
        loanType:
          $ref: "#/components/schemas/LoanType"
        lenderName:
          type: string
        totalLoanAmount:
          type: number
        outstandingAmount:
          type: number
        interestRate:
          type: number
        emi:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        familyMemberId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
    Liability:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        familyMemberId:
          type: integer
          nullable: true
        loanType:
          $ref: "#/components/schemas/LoanType"
        lenderName:
          type: string
        totalLoanAmount:
          type: number
        outstandingAmount:
          type: number
        interestRate:
          type: number
        emi:
          type: number
        startDate:
          type: string
          nullable: true
        endDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
          description: Additional details or specific category breakdown
        createdAt:
          type: string
      required:
        - id
        - clientId
        - loanType
        - lenderName
        - totalLoanAmount
        - outstandingAmount
        - interestRate
        - emi
        - createdAt
...